requirements.txt
--extra-index-url=http://dist.pinaxproject.com/dev/
--extra-index-url=http://dist.pinaxproject.com/alpha/

Django==1.4.1
django-compressor
pinax-theme-bootstrap==2.0.4
pinax-theme-bootstrap-account==1.0b2
django-user-accounts==1.0b3
django-forms-bootstrap==2.0.3.post1
django-crispy-forms
-e git+https://github.com/yourlabs/django-zodb-light.git#egg=zlight
-e git+https://github.com/jpic/zodb-admin.git#egg=admin
-e git+https://github.com/yourlabs/django-appstore.git#egg=admin
django-annoying
django-session-security
pinax-utils==1.0b1.dev3
django-debug-toolbar
django-autocomplete-light
-e git+git://github.com/gkuhn1/django-admin-templates-twitter-bootstrap.git#egg=admin-bootstrap
